float error

+++ real_estate_advert/paruvendu/parser.py
@@ -28,18 +28,18 @@ def ParseParuvendu(data):
     if "Location" in adtyp:adtyp = "rent"
     else:adtyp="buy"
     area = data.get("title")[data.get("title").rfind("- ")+2:].replace("m²","").strip()
-    try:area = float(area)
-    except: area = float(detailV.get("SUR")) if detailV.get("SUR") else 0
+    try:area = int(area)
+    except: area = int(detailV.get("SUR")) if detailV.get("SUR") else 0
 
     pinRegx = r'(\d{5}\-?\d{0,4})'
     sdata = {
       "id": data.get("id"),
       "ads_type": adtyp,
-      "price": float(data.get("price").replace(" ",'')),
-      "original_price": float(data.get("price").replace(" ",'')),
+      "price": int(data.get("price").replace(" ",'')),
+      "original_price": int(data.get("price").replace(" ",'')),
       "area": area ,
       "city": data.get("subtitle")[:data.get("subtitle").find(" (")],     # "Poitiers (86000)"
-      "declared_habitable_surface": float(detailV.get("SUR")) if detailV.get("SUR") else 0,
+      "declared_habitable_surface": int(detailV.get("SUR")) if detailV.get("SUR") else 0,
       "declared_land_surface": detailT.get("9999999_125"),
       "land_surface": detailT.get("9999999_125"),
       "declared_rooms": detailT.get("NBP"),
